+++ src/contents/sounds.tsx
@@ -1,7 +1,10 @@
 import WindowWrapper from "@/components/window";
 import { useTranslation } from "react-i18next";
 import { sounds } from "@/lib/sounds";
-import SoundCard from "@/components/sound-card";
+import { lazy, Suspense } from "react";
+import { SoundCardLoader } from "@/loaders/sounds";
+
+const SoundCard = lazy(()=>import("@/components/sound-card"));
 
 export default function SoundsContent(){
      const {t} = useTranslation("sounds")
@@ -9,7 +12,9 @@ export default function SoundsContent(){
           <WindowWrapper title={t("title")} className="mt-8 mb-12 flex justify-center items-center flex-col gap-2 w-full">
                <div className="w-full grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-1.5">
                     {sounds.map(sound=>(
-                         <SoundCard key={sound.id} data={sound}/>
+                         <Suspense key={sound.id} fallback={<SoundCardLoader/>}>
+                              <SoundCard data={sound}/>
+                         </Suspense>
                     ))}
                </div>
           </WindowWrapper>