+++ src/lib/sounds.ts
@@ -1,5 +1,17 @@
+import { lazy } from "react"
 import { ISounds } from "./types"
-import {Birds, Rain, River, Beach, Train, Thunderstorm, Night, Lake, Waves, Waterfall} from "@/icons"
+import {} from "@/icons"
+
+const Birds = lazy(()=>import("@/icons/birds"))
+const Rain = lazy(()=>import("@/icons/rain"))
+const River = lazy(()=>import("@/icons/river"))
+const Beach = lazy(()=>import("@/icons/beach"))
+const Train = lazy(()=>import("@/icons/train"))
+const Thunderstorm = lazy(()=>import("@/icons/thunderstorm"))
+const Night = lazy(()=>import("@/icons/night"))
+const Lake = lazy(()=>import("@/icons/lake"))
+const Waves = lazy(()=>import("@/icons/waves"))
+const Waterfall = lazy(()=>import("@/icons/waterfall"))
 
 export const soundIds = [
      "birds", "rain",