+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 ## [0.4.0] - Unreleased
+### Added
+- Skeleton loading on Icons and Logo Dropdown
+- About Page Content with 3 links
 ### Improved
 - Sound Card UI
 - Sounds page loader